refactor: adopt Rstack CLI for hooks, builds, tests, and docs
- Replace simple-git-hooks and nano-staged with rs hooks and rs staged.
- Build all nine packages with rs lib via define.lib(), retaining addon-rslib's @rslib/core dev copy for its peer.
- Run project tests through rs test with define.test() and rstest.shared.ts.
- Run the website through rs doc and update the tooling docs.
