+++ .rstack/hooks/pre-commit
@@ -1,0 +1,1 @@
+rstack staged
