+++ docs/dependencies.md
@@ -1,9 +1,9 @@
 # Dependencies
 
 - After editing any dependency manifest (`package.json`, `pnpm-workspace.yaml`, `.npmrc`), run `pnpm install` immediately.
-- After a `rebase`, `merge`, or `stash pop` that changed dependency files (including `pnpm-lock.yaml`), run `pnpm install` before any validation command (`build`, `test`, `e2e`, `lint`) — stale `node_modules` produces failures that look like regressions.
+- After a `rebase`, `merge`, or `stash pop` that changed dependency files (including `pnpm-lock.yaml`), run `pnpm install` before any validation command (`build`, `test`, `e2e`, `check`) — stale `node_modules` produces failures that look like regressions.
 - Commit the resulting `pnpm-lock.yaml` changes in the same branch. Never push dependency manifest changes without the matching `pnpm-lock.yaml` update.
-- `storybook-rsbuild-scripts` is intentionally a root-only `devDependency`: package-local Rslib configs resolve it from the root `node_modules`, keeping this private workspace package out of published package manifests.
+- `storybook-rsbuild-scripts` is intentionally a root-only `devDependency`: package-local `rstack.config.ts` files resolve it from the root `node_modules` when registering their Rslib config with `define.lib()`, keeping this private workspace package out of published package manifests.
 - In `packages/*`, dependency type is a bundle-vs-external switch, not dev-vs-prod: `createRslibConfig` enables `output.autoExternal`, which externalizes `dependencies`, `optionalDependencies`, and `peerDependencies` while leaving `devDependencies` bundled; each package's self-reference is an explicit external. Some runtime imports (e.g. `slash`, `pretty-hrtime` in `storybook-builder-rsbuild`) sit in `devDependencies` deliberately so they get vendored into the bundle — "fixing" them into `dependencies` changes what is published. Check the factory's external policy before reclassifying a dependency.
 - `peerDependencies` are hand-maintained: Renovate's global rule disables `peerDependencies` updates, except the `storybook` group in `.github/renovate.json5`, which bumps `dependencies`, `devDependencies`, and `peerDependencies` for `storybook` / `@storybook/**` together in one PR.
 - Non-Storybook peer ranges (`@rsbuild/core`, `react`, `react-dom`) are deliberately **wider** than the matching `devDependencies`. Never narrow them: `pnpm check-dependency-version` only cross-checks that peer ranges agree between packages, so a narrowed range stays green while silently dropping supported consumers.
