+++ docs/testing.md
@@ -4,7 +4,7 @@
 
 - Tests live in the package-level `tests/` directory, mirroring the `src/` structure (e.g. `packages/builder-rsbuild/tests/preview/`). Cross-package integration tests live in the root `tests/` project.
 - Exception: the ported `packages/framework-react/src/plugins/react-docgen-typescript/__tests__/` suite (with its `__fixtures__`/`__snapshots__`) stays in its upstream layout — never relocate it.
-- The root `rstest.config.ts` aggregates projects by glob — `./packages/*/rstest.config.ts`, `./sandboxes/*/rstest.config.ts`, and `./tests/rstest.config.ts`. A package or sandbox without its own `rstest.config.ts` is silently skipped, with no error.
+- The root `rstack.config.ts` registers the Rstest projects with `define.test()` using the globs `./packages/*/rstest.config.ts`, `./sandboxes/*/rstest.config.ts`, and `./tests/rstest.config.ts`. Shared project defaults live in the plain `rstest.shared.ts` module. A package or sandbox without its own `rstest.config.ts` is silently skipped, with no error.
 - The root `tests/` project sets `include: ['./*.test.ts']` — top level only. A file at `tests/<subdir>/foo.test.ts` never runs and never warns; put cross-package integration tests directly in `tests/` and keep shared code in `tests/helpers/` (not a test glob).
 - The shared `rstest-setup.ts` replaces `console.warn`/`console.error` with functions that **throw** unless the message matches its `ignoreList` — a code path that warns fails as a bare `Error: warn: ...` with a stack pointing into `src/`. Spy on the console in the test instead; never silence the warning in `src/`. (The root `tests/` project uses its own spy-free setup.)
 - `pnpm test` is not self-contained: the chromatic suites read `sandboxes/react-18/storybook-static/preview-stats.json`, a gitignored build output. On a clean checkout run `pnpm build:sandboxes` first (CI's order) — those `ENOENT` failures are not a regression in your change.
