+++ packages/builder-rsbuild/rstack.config.ts
@@ -1,0 +1,15 @@
+import { define } from 'rstack'
+import { createRslibConfig } from 'storybook-rsbuild-scripts/create-rslib-config'
+import buildConfig from './build-config.ts'
+
+define.lib(
+  createRslibConfig(import.meta.dirname, buildConfig, {
+    define:
+      process.env.SB_RSBUILD_TEST_MINIMAL_DEV === 'true'
+        ? {
+            'process.env.SB_RSBUILD_TEST_MINIMAL_DEV': JSON.stringify('true'),
+          }
+        : {},
+    disableUrlParsingFor: /builder-rsbuild[\\/]compiled/,
+  }),
+)
