+++ packages/framework-html/rstack.config.ts
@@ -1,6 +1,5 @@
+import { define } from 'rstack'
 import { createRslibConfig } from 'storybook-rsbuild-scripts/create-rslib-config'
 import buildConfig from './build-config.ts'
 
-export default createRslibConfig(import.meta.dirname, buildConfig, {
-  dtsBundleTsconfigPath: './tsconfig.dts-bundle.json',
-})
+define.lib(createRslibConfig(import.meta.dirname, buildConfig))
