+++ packages/framework-react-native-web/package.json
@@ -41,8 +41,8 @@
     "!src/**/*"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build --watch",
+    "build": "rstack lib",
+    "dev": "rstack lib --watch",
     "prepare": "pnpm run build",
     "type-check": "jiti ../../scripts/check/check-package.ts"
   },
@@ -55,7 +55,6 @@
   "devDependencies": {
     "@microsoft/api-extractor": "^7.59.0",
     "@rsbuild/core": "^2.1.13",
-    "@rslib/core": "~1.0.0-rc.2",
     "@types/node": "^22.0.0",
     "@types/react": "^18.3.31",
     "react": "18.3.1",
