+++ packages/framework-vue3/package.json
@@ -39,8 +39,8 @@
     "!src/**/*"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build --watch",
+    "build": "rstack lib",
+    "dev": "rstack lib --watch",
     "prepare": "pnpm run build",
     "type-check": "jiti ../../scripts/check/check-package.ts"
   },
@@ -53,7 +53,6 @@
   "devDependencies": {
     "@microsoft/api-extractor": "^7.59.0",
     "@rsbuild/core": "^2.1.13",
-    "@rslib/core": "~1.0.0-rc.2",
     "@types/node": "^22.0.0",
     "storybook": "10.5.10",
     "typescript": "^5.9.3",
