--- packages/framework-vue3/rslib.config.ts
@@ -1,4 +1,0 @@
-import { createRslibConfig } from 'storybook-rsbuild-scripts/create-rslib-config'
-import buildConfig from './build-config.ts'
-
-export default createRslibConfig(import.meta.dirname, buildConfig)
