+++ packages/rsbuild-plugin-react-native-web/rstack.config.ts
@@ -1,0 +1,9 @@
+import { define } from 'rstack'
+import { createRslibConfig } from 'storybook-rsbuild-scripts/create-rslib-config'
+import buildConfig from './build-config.ts'
+
+define.lib(
+  createRslibConfig(import.meta.dirname, buildConfig, {
+    dtsBundleTsconfigPath: './tsconfig.dts-bundle.json',
+  }),
+)
