+++ sandboxes/react-18/rstest.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from '@rstest/core'
-import { rstestCommonConfig } from '../../rstest.config'
+import { rstestCommonConfig } from '../../rstest.shared.ts'
 
 export default defineConfig({
   ...rstestCommonConfig,
